Added Behavior Tree structure

+++ sw/airborne/firmwares/rotorcraft/autopilot.c
@@ -40,6 +40,8 @@
 #include "firmwares/rotorcraft/guidance.h"
 #include "firmwares/rotorcraft/stabilization.h"
 
+#include "modules/sensors/RSSI2dist.h"
+
 #ifdef POWER_SWITCH_GPIO
 #include "mcu_periph/gpio.h"
 #endif
@@ -180,8 +182,8 @@ static void send_fp(void) {
 
 static void send_bluegps(void)
 {
-  int16_t dEst = (int16_t)(rssiEkf_estimate.v[0]*1000);
-  int16_t vEst = (int16_t)(rssiEkf_estimate.v[1]*1000);
+  int32_t dEst = (int32_t)(rssiEkf_estimate.v[0]*1000);
+  int32_t vEst = (int32_t)(rssiEkf_estimate.v[1]*1000);
   DOWNLINK_SEND_BLUEGPS(DefaultChannel, DefaultDevice,
                               &(rssi[0]),
                               &(stateGetPositionEnu_i()->x),
